Extract dialog creation to a collaborator.
