+++ demo/src/main/java/com/novoda/demo/MainActivity.java
@@ -1,27 +1,16 @@
 package com.novoda.demo;
 
 import android.app.Activity;
-import android.app.AlertDialog;
-import android.content.DialogInterface;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.ArrayAdapter;
 import android.widget.Toast;
 
 import com.novoda.noplayer.ContentType;
 import com.novoda.noplayer.NoPlayer;
 import com.novoda.noplayer.PlayerBuilder;
 import com.novoda.noplayer.PlayerView;
-import com.novoda.noplayer.model.AudioTracks;
-import com.novoda.noplayer.model.PlayerAudioTrack;
-import com.novoda.noplayer.model.PlayerSubtitleTrack;
-import com.novoda.noplayer.model.PlayerVideoTrack;
 import com.novoda.utils.NoPlayerLog;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Locale;
 
 public class MainActivity extends Activity {
 
@@ -33,6 +22,7 @@ public class MainActivity extends Activity {
 
     private NoPlayer player;
     private DemoPresenter demoPresenter;
+    private DialogCreator dialogCreator;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -57,6 +47,7 @@ public class MainActivity extends Activity {
                 .build(this);
 
         demoPresenter = new DemoPresenter(controllerView, player, player.getListeners(), playerView);
+        dialogCreator = new DialogCreator(this, player);
     }
 
     @Override
@@ -73,32 +64,8 @@ public class MainActivity extends Activity {
             if (player.getVideoTracks().isEmpty()) {
                 Toast.makeText(MainActivity.this, "no additional video tracks available!", Toast.LENGTH_LONG).show();
             } else {
-                showVideoSelectionDialog();
+                dialogCreator.showVideoSelectionDialog();
             }
-        }
-
-        private void showVideoSelectionDialog() {
-            final List<PlayerVideoTrack> videoTracks = player.getVideoTracks();
-            ArrayAdapter<String> adapter = new ArrayAdapter<>(MainActivity.this, R.layout.list_item);
-            adapter.addAll(mapVideoTrackToLabel(videoTracks));
-            AlertDialog videoTrackSelectionDialog = new AlertDialog.Builder(MainActivity.this)
-                    .setTitle("Select Video track")
-                    .setAdapter(adapter, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int position) {
-                            PlayerVideoTrack videoTrack = videoTracks.get(position);
-                            player.selectVideoTrack(videoTrack);
-                        }
-                    }).create();
-            videoTrackSelectionDialog.show();
-        }
-
-        private List<String> mapVideoTrackToLabel(List<PlayerVideoTrack> videoTracks) {
-            List<String> labels = new ArrayList<>();
-            for (PlayerVideoTrack videoTrack : videoTracks) {
-                labels.add("Quality " + videoTrack.height());
-            }
-            return labels;
         }
     };
 
@@ -106,38 +73,7 @@ public class MainActivity extends Activity {
 
         @Override
         public void onClick(View v) {
-            showAudioSelectionDialog();
-        }
-
-        private void showAudioSelectionDialog() {
-            final AudioTracks audioTracks = player.getAudioTracks();
-            ArrayAdapter<String> adapter = new ArrayAdapter<>(MainActivity.this, R.layout.list_item);
-            adapter.addAll(mapAudioTrackToLabel(audioTracks));
-            AlertDialog audioSelectionDialog = new AlertDialog.Builder(MainActivity.this)
-                    .setTitle("Select audio track")
-                    .setAdapter(adapter, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int position) {
-                            PlayerAudioTrack audioTrack = audioTracks.get(position);
-                            player.selectAudioTrack(audioTrack);
-                        }
-                    }).create();
-            audioSelectionDialog.show();
-        }
-
-        private List<String> mapAudioTrackToLabel(AudioTracks audioTracks) {
-            List<String> labels = new ArrayList<>();
-            for (PlayerAudioTrack audioTrack : audioTracks) {
-                String label = String.format(
-                        Locale.UK,
-                        "Group: %s Format: %s Type: %s",
-                        audioTrack.groupIndex(),
-                        audioTrack.formatIndex(),
-                        audioTrack.audioTrackType()
-                );
-                labels.add(label);
-            }
-            return labels;
+            dialogCreator.showAudioSelectionDialog();
         }
     };
 
@@ -145,42 +81,10 @@ public class MainActivity extends Activity {
         @Override
         public void onClick(View v) {
             if (!player.getSubtitleTracks().isEmpty()) {
-                showSubtitleSelectionDialog();
+                dialogCreator.showSubtitleSelectionDialog();
             } else {
                 Toast.makeText(MainActivity.this, "no subtitles available!", Toast.LENGTH_LONG).show();
             }
-        }
-
-        private void showSubtitleSelectionDialog() {
-            final List<PlayerSubtitleTrack> subtitleTracks = player.getSubtitleTracks();
-            ArrayAdapter<String> adapter = new ArrayAdapter<>(MainActivity.this, R.layout.list_item);
-            adapter.addAll(mapSubtitleTrackToLabel(subtitleTracks));
-            AlertDialog subtitlesSelectionDialog = new AlertDialog.Builder(MainActivity.this)
-                    .setTitle("Select subtitle track")
-                    .setAdapter(adapter, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int position) {
-                            switch (position) {
-                                case 0:
-                                    player.hideSubtitleTrack();
-                                    break;
-                                default:
-                                    PlayerSubtitleTrack subtitleTrack = subtitleTracks.get(position - 1);
-                                    player.showSubtitleTrack(subtitleTrack);
-                                    break;
-                            }
-                        }
-                    }).create();
-            subtitlesSelectionDialog.show();
-        }
-
-        private List<String> mapSubtitleTrackToLabel(List<PlayerSubtitleTrack> subtitleTracks) {
-            List<String> labels = new ArrayList<>();
-            labels.add("Dismiss subtitles");
-            for (PlayerSubtitleTrack subtitleTrack : subtitleTracks) {
-                labels.add("Group: " + subtitleTrack.groupIndex() + " Format: " + subtitleTrack.formatIndex());
-            }
-            return labels;
         }
     };
